+++ include/stats.hpp
@@ -9,16 +9,16 @@ namespace parameters
 		size_t t = 0;
 		size_t evaluations = 0;
 
-		double current_avg = std::numeric_limits<double>::infinity();
+		Float current_avg = std::numeric_limits<Float>::infinity();
 		std::vector<Solution> solutions = {};
 		std::vector<Solution> centers = {};
 		Solution current_best = {};
 		Solution global_best = {};
 		bool has_improved = false;
-		double success_ratio = 2.0 / 11.0;
-		double cs = 1.0 / 12.0;
+		Float success_ratio = 2.0 / 11.0;
+		Float cs = 1.0 / 12.0;
 
-		void update_best(const Vector &x, const double y)
+		void update_best(const Vector &x, const Float y)
 		{
 			has_improved = false;
 			if (y < current_best.y)
